// Utils/ENUM.ts
export enum accountCredientials {
    USERNAME = 'Test',
    PASSWORD = 'pwd',

}

export enum messages {
    SUCCESFULL = 'Welcome, ',
    UNSUCCESFULL = 'Invalid username/password',
    LOGOUT = 'User logged out.',
}

export enum overlappedValues {
    ID = '3456',
    NAME = 'Faruk SEVIL',
}

export enum api {
    BASE_URL = 'https://petstore.swagger.io/v2',
}